+++ Utilities/Crypto/Hashes.hpp
@@ -308,7 +308,7 @@ namespace Hash
         unsigned int Buffersize = 20;
         unsigned char Buffer[20]{};
 
-        HMAC(EVP_sha1(), Key, Keysize, (uint8_t *)Input, Size, Buffer, &Buffersize);
+        HMAC(EVP_sha1(), Key, (int)Keysize, (uint8_t *)Input, (int)Size, Buffer, &Buffersize);
         return std::string((char *)Buffer, Buffersize);
     }
     inline std::string HMACSHA256(const void *Input, const size_t Size, const void *Key, const size_t Keysize)
@@ -316,7 +316,7 @@ namespace Hash
         unsigned int Buffersize = 32;
         unsigned char Buffer[32]{};
 
-        HMAC(EVP_sha256(), Key, Keysize, (uint8_t *)Input, Size, Buffer, &Buffersize);
+        HMAC(EVP_sha256(), Key, (int)Keysize, (uint8_t *)Input, (int)Size, Buffer, &Buffersize);
         return std::string((char *)Buffer, Buffersize);
     }
     inline std::string HMACSHA512(const void *Input, const size_t Size, const void *Key, const size_t Keysize)
@@ -324,7 +324,7 @@ namespace Hash
         unsigned int Buffersize = 64;
         unsigned char Buffer[64]{};
 
-        HMAC(EVP_sha512(), Key, Keysize, (uint8_t *)Input, Size, Buffer, &Buffersize);
+        HMAC(EVP_sha512(), Key, (int)Keysize, (uint8_t *)Input, (int)Size, Buffer, &Buffersize);
         return std::string((char *)Buffer, Buffersize);
     }
     #endif
